airnb_backend/requirements.txt
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
cryptography==42.0.5
filelock==3.14.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
SQLAlchemy==2.0.29
stripe==9.4.0
tldextract==5.1.2
typing_extensions==4.11.0
urllib3==2.2.1
watchdog==4.0.0
zope.event==5.0
zope.interface==6.3
